requirements.txt
customtkinter>=5.2.2
SpeechRecognition>=3.10.1
PyAudio>=0.2.14
pyttsx3>=2.90
google-generativeai>=0.7.2
